vignette now builds ok
What seems to be happening is that `plot()` is calling the base
function, not the specialization for an oce/ctd object. If I change the
doc to use e.g. `oce::plot()` on a ctd object, then all works fine.

I didn't think this was necessary. I thought R would check the signature
of an object, see it was "glider" and not "ctd", and dispatch
appropriately.

This closes #9. Although I will still keep thinking about this, the
issue as stated in the title has been addressed, so I will close this.