summary: revert back to order as in file
We have an issue on this now. I had thought ordering at summary time
would be nice, but users might think the first listed could be accessed
as [[1]], etc., so that's bad.  Another alternative would be to do the
ordering at read time, and to handle insertion done by function calls.
But what if a user inserts *without* function calls ... then the order
will go wherever it goes. And what about renaming? (Well, frankly, if
users do that, they will open up a host of problems, because flag names
and unit names are linked to the nams assigned during reading.)