isolate plot...
This is to help me see why the plot does not upate until I change the
radio button for the plot type. I see that the first plot *does* get
created ... but it's created in the Rstudio plot pane, not in the shiny
output.  Maybe this a bug in shiny or rstudio.