edit system roughed in
But I don't like that the 'bad' vector is the length of the data --
pretty soon the edits will make the rda files double in size!  Maybe we
ought to have

list(time, index, old, new)

and only save indices that are changed.  This will be useful in an undo
scheme, also.