add tmp debugging
This was for https://github.com/dankelley/oceanglider/issues/54 but,
mysteriously, the problem went away when I put in these lines. Maybe
RStudio was freaked out, somehow.  (It *was* kind of messed up because I
noticed it was not reloading qc02.R when I edited it separately.)

I will check this on my home machine before spending much time thinking
about it.