ability to flag whole yo; also:
* I think I have the yo specification working now, using throttle()
instead of debounce().

* I removed the next- and previous-yo buttons, since the little arrows
seem to work fine now, with throttle() being used.

* I added some more notifications for slow operations. I may try doing
that for plots, also (have to check whether it forces a redraw after the
dialog gets removed though ... that would defeat the purpose!)

* Likely some other things. Just checking in because I emailed CR with a
question about air gulps, and wanted the code to work in case he wants
to try.