I think debouncing works...
... although I've said that before!

I also have a new function to calculate indices of data being displayed
currently. I will need that for brushing and hovering.

Also, I have a UI element that permits setting the brush response. This
is because I was thinking of making brushing work for highlighting data,
but I am now not sure on that.

I am not sure what clicking does at the moment, or what it should do. As
a user, I find it confusing to have both click and brush.