powerUp: start coding but
it is really far too slow to work. I think I need to debounce the
slider, and also I ought to have an observer, so that the slow work gets
done and saved in a global variable to be used by visibleIndices().

Maybe some of the other work in visibleIndices() is also too slow.