Prepare for use of updated oce
In the "develop" github branch of Oce, handleFlags() has a new argument,
`where`, which means that oceanglider handleFlags() can be simplified
quite a lot.  This commit does that.

The tests in tests/testthat/test_flags.R ought to be sufficient to show
that doing this has not broken anything.

Simplifying oceanglider in this way should yield higher robustness,
because the oce::handleFlags() has a large user base for error
detection.