rewrite all docs in markdown
I put the following in my ~/.vimrc, to help with this.

   :map <F1> :s/\\code{\([^}]*\)}/`\1`/gc

My procedure is to highlight a block of text, press function-1, and then
enter y/n on each item.  This does not fix instances of \code{\link{}}
but those are easy to translate to e.g. `function()` for a function in
the package, to `item` for a non-functoin part of a package, or for e.g.
`oce::function()` or `oce::item()` for things in the oce package.