improve colorizing by 'N2 extremes'
but I think something is wrong for profile plots, e.g. with g21m51 a
density profile shows red dots not at the bottom of the water column.

Either I'm plotting wrong or maybe my N2 is overly smoothed?? (I am
pretty sure it is a plotting -- that is, a subsetting -- problem.)