can now load e.g. sea021m51_(TEXT).rda
This may be handy for users who want to name files. Previously, the
(TEXT) part had to be a date.

I may add something to the 'save' action that asks the users for such
text, so they don't have to go into the OS to rename files. Actually,
maybe I should also ask them for free-form text, in which they might
provide notes on the procedure.