*experimental* plot colorBy: lnN2
I am not sure why there are so many strongly unstable points in the
mixed layer.  Maybe my N2 algorithm is not good.

NOTE: this only works for profile plots ... for others, it gives errors.
This is because I see no point in propogating th method to 3 spots in
the code, unless or until the algorithm makes sense to me.

I may end up dropping this.