new instability calculation, in qc08.R
NOTE: qc08.R is the new standard. It now calculates gravitational
instability, and you can colour-code for it by setting the colour scheme
to 'instability'. Since this new signal is stored in the outgoing glider
object, it is not recommended to use .rda files created with qc07.R as
"previous analysis", and I recommend deleting such files, to avoid
confusion.  (Actually, everything will work, except for colourizing by
instability, which might yield an error.)