drawPalette(cex=) reframed as (cex.palette=)
This is to fit with the new oce behaviour.  It will cause errors with
people who do not update their oce package, but since the oceanglider
repo is currently aimed at experts, that ought not to be a problem,
and the benefit is that we get a clean build/check.