move plot,glider-method to plotGlider()
I thought what I had would work, and maybe it did in the past, but
it fails now.  I just don't know the proper way to make a generic plot
function for this.  It works fine in the console but fails in a
vignette.  I think the problem is something to do with inheritance from
oce. Or not.  At this stage I don't care what the problem was, because
the user should be able to either remember to use plotGlider() or to do
a ??oceanglider if they try plot() and it fails. Dedicating time to
tracking down this problem is unwise, given the existence of more
compelling tasks.