reformat code (and possibly fix oxygen calib issue)
The issue was not reported.  But, lintr said I was not using a variable
and indeed I was not.  This was for a temp/degC vs temp/K mixup.  I
haven't looked at the coefficients to see whether this will make a small
difference or a large one.

At some point, it would be great to look into this some more, and then
to add some checks to the test suite, so we can have more confidence in
the oxygen values.