rewrite flag functions
After several wasted hours, I have given up on deriving the flag-related
generic methods from the oce package.

Now, users are to employ e.g. handleGliderFlags() instead of
handleFlags(), the latter being the oce method.  This should not be hard
remember, and orememberf course the functions are documented.

The problem of inheriting generic functions is not particular to
oceglider; we had somewhat similar issues with argoFloats.  I think
there must be something wrong with my understanding of object
inheritance in R, as opposed to other languages.

I may revisit this later.